split inverted_index.rs into submodule files
Cleaner structure and will reduce the risk of conflicts when working on
multiple iterators in parallel.
