+++ src/redisearch_rs/rqe_iterators/src/inverted_index/mod.rs
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) 2006-Present, Redis Ltd.
+ * All rights reserved.
+ *
+ * Licensed under your choice of the Redis Source Available License 2.0
+ * (RSALv2); or (b) the Server Side Public License v1 (SSPLv1); or (c) the
+ * GNU Affero General Public License v3 (AGPLv3).
+*/
+
+//! Supporting types for [`Numeric`], [`Term`], and [`Wildcard`].
+
+mod core;
+mod numeric;
+mod term;
+mod wildcard;
+
+pub use numeric::Numeric;
+pub use term::Term;
+pub use wildcard::Wildcard;
