+++ src/redisearch_rs/rqe_iterators/src/inverted_index/term.rs
@@ -1,0 +1,125 @@
+/*
+ * Copyright (c) 2006-Present, Redis Ltd.
+ * All rights reserved.
+ *
+ * Licensed under your choice of the Redis Source Available License 2.0
+ * (RSALv2); or (b) the Server Side Public License v1 (SSPLv1); or (c) the
+ * GNU Affero General Public License v3 (AGPLv3).
+*/
+
+use std::ptr::NonNull;
+
+use ffi::{RS_FIELDMASK_ALL, RedisSearchCtx, t_docId};
+use inverted_index::{RSIndexResult, RSOffsetSlice, TermReader};
+use query_term::RSQueryTerm;
+
+use crate::{
+    RQEIterator, RQEIteratorError, RQEValidateStatus, SkipToOutcome,
+    expiration_checker::ExpirationChecker,
+};
+
+use super::core::InvIndIterator;
+
+/// An iterator over term inverted index entries.
+///
+/// This iterator can be used to query a term inverted index.
+///
+/// # Type Parameters
+///
+/// * `'index` - The lifetime of the index being iterated over.
+/// * `R` - The reader type used to read the inverted index.
+/// * `E` - The expiration checker type used to check for expired documents.
+pub struct Term<'index, R, E = crate::expiration_checker::NoOpChecker> {
+    it: InvIndIterator<'index, R, E>,
+    #[allow(dead_code)] // will be used by should_abort()
+    context: NonNull<RedisSearchCtx>,
+}
+
+impl<'index, R, E> Term<'index, R, E>
+where
+    R: TermReader<'index>,
+    E: ExpirationChecker,
+{
+    /// Create an iterator returning results from a term inverted index.
+    ///
+    /// Filtering the results can be achieved by wrapping the reader with
+    /// a [`inverted_index::FilterMaskReader`].
+    ///
+    /// `term` is the query term that brought up this iterator. It is stored
+    /// in the result and persists across all reads.
+    ///
+    /// `weight` is the scoring weight applied to the result record. It is
+    /// typically derived from the query node and used by the scoring function
+    /// to scale the relevance of results from this iterator.
+    ///
+    /// `expiration_checker` is used to check for expired documents when reading from the inverted index.
+    ///
+    /// # Safety
+    ///
+    /// 1. `context` must point to a valid [`RedisSearchCtx`].
+    /// 2. `context` must remain valid for the lifetime of the iterator.
+    pub unsafe fn new(
+        reader: R,
+        context: NonNull<RedisSearchCtx>,
+        term: Box<RSQueryTerm>,
+        weight: f64,
+        expiration_checker: E,
+    ) -> Self {
+        let result =
+            RSIndexResult::with_term(Some(term), RSOffsetSlice::empty(), 0, RS_FIELDMASK_ALL, 1)
+                .weight(weight);
+        Self {
+            it: InvIndIterator::new(reader, result, expiration_checker),
+            context,
+        }
+    }
+}
+
+impl<'index, R, E> RQEIterator<'index> for Term<'index, R, E>
+where
+    R: TermReader<'index>,
+    E: ExpirationChecker,
+{
+    #[inline(always)]
+    fn current(&mut self) -> Option<&mut RSIndexResult<'index>> {
+        self.it.current()
+    }
+
+    #[inline(always)]
+    fn read(&mut self) -> Result<Option<&mut RSIndexResult<'index>>, RQEIteratorError> {
+        self.it.read()
+    }
+
+    #[inline(always)]
+    fn skip_to(
+        &mut self,
+        doc_id: t_docId,
+    ) -> Result<Option<SkipToOutcome<'_, 'index>>, RQEIteratorError> {
+        self.it.skip_to(doc_id)
+    }
+
+    #[inline(always)]
+    fn rewind(&mut self) {
+        self.it.rewind()
+    }
+
+    #[inline(always)]
+    fn num_estimated(&self) -> usize {
+        self.it.num_estimated()
+    }
+
+    #[inline(always)]
+    fn last_doc_id(&self) -> t_docId {
+        self.it.last_doc_id()
+    }
+
+    #[inline(always)]
+    fn at_eof(&self) -> bool {
+        self.it.at_eof()
+    }
+
+    #[inline(always)]
+    fn revalidate(&mut self) -> Result<RQEValidateStatus<'_, 'index>, RQEIteratorError> {
+        self.it.revalidate()
+    }
+}
